fix(review): address review feedback and static analysis findings

+++ GenHub/GenHub.Core/Services/Tools/Checksum/BigArchiveEntry.cs
@@ -7,4 +7,4 @@ namespace GenHub.Core.Services.Tools.Checksum;
 /// <param name="ArchivePath">The absolute path of the containing .BIG archive.</param>
 /// <param name="Offset">The byte offset of the entry data inside the archive.</param>
 /// <param name="Size">The byte size of the entry data.</param>
-public sealed record BigArchiveEntry(string Path, string ArchivePath, int Offset, int Size);
+public sealed record BigArchiveEntry(string Path, string ArchivePath, long Offset, long Size);
